requirements.txt
yfinance
pandas
ta
google-api-python-client
google-auth
google-auth-oauthlib